comments page is constructed according to new response from server
